sharing [nfc]: Construct stream/topic narrows only for real streams
This way the narrows we're making are valid; and this code will be
ready to straightforwardly start supply stream IDs in constructing
these narrows, when we start accepting those.

This change is NFC because both of the places we pass these narrows
-- fetchTopicsForStream, and TopicAutocomplete -- will already do
nothing if the stream name in the narrow is one that doesn't exist
in our data.