notif: Find stream on opening a stream-message notification
In a notification for a stream message, the server doesn't send us
the stream's ID, only its name.  We'll soon want to use stream IDs
in our Narrow values, which means we'll need the stream's ID up
front when the user opens such a notification.  So, start looking
up the stream in our data structures at this point.

There's an open issue for the server to start sending stream IDs here:
  https://github.com/zulip/zulip/issues/18067
When it does, we can start looking for that and using it when present,
but we'll want to keep this lookup as a fallback for a while after.

This change causes getNarrowFromNotificationData to start returning
null if the notification is for a stream we don't know about, where
previously it would go ahead and return a narrow.  The old behavior
is that we'd navigate to that narrow, and it would show an error
message "That conversation doesn't seem to exist" (via InvalidNarrow).
The new behavior is that we just won't act on the notification at
all.  That's less good, but tolerable, and what we already do for
1:1 PMs; leave a TODO comment explaining it.