compose: Give a more helpful alert when camera is unavailable.
`errorMessage` is documented as for debugging only [1]. So, give our
own simple message for this case, and also make sure it's set up for
translation.

We do offer the Android app on the Play Store for devices that don't
have cameras, since #4724, and we don't hide the camera button on
those devices.

[1]  https://github.com/react-native-image-picker/react-native-image-picker#the-response-object