msglist: Add new div#msglist-elements
For our upcoming message-list diffing algorithm, for #3851, the code
outside the WebView will be sending edit instructions into the
WebView.

Those instructions will include indexes into the elements list, like
"go add message A at index i". At the time the in-WebView code reads
those instructions and decides to act on them, it'll be really
important that that inner code agrees with the code outside the
WebView about what elements are at which indexes.

We'll want to think carefully about how to answer that requirement
correctly, avoiding sneaky races, etc.

For now, though, just give the inside-the-WebView code a reasonable
way to work with indexes in the first place: the nth child of this
new div#msglist-elements is index n in the list.

And with the goal of being not disruptive, adjust some code that
looked like it was relying on the DOM structure before this change,
so that it works after it.