AppEventHandlers [nfc]: Mark a chaotic use of per-account state
This global component is using per-account state, from the active
account, to control some global UI -- the badge count on the app's
icon in the launcher, for some Android varieties that have such a
thing.

The mismatch of types actually highlights how the feature is pretty
broken if you use multiple accounts, in that the count reflects only
the one last account you used.  Perhaps we should just cut the
feature.

In any case, for now, mark the type mismatch so that we can tighten up
these types in general.