api [nfc]: Cut false flexibility in register-queue binding
We'd been having this function take parameters from its caller
to say what kinds of data it should request, and what format it's
prepared to handle the data in.

But in fact the answers to those questions live inside this `src/api/`
API-bindings subtree, and some of them in this very file:

 * In `initialDataTypes.js` we describe what properties we expect to
   get in the response to this request, which corresponds to what
   event types we mention in `fetch_event_types` here.

 * In the "transform" helpers in this file (and in the `AvatarURL`
   class which they invoke to do the work), we take care of the
   client-side functionality that `user_avatar_url_field_optional`
   and `client_gravatar` tell the server we have.

 * The remaining client capabilities `notification_settings_null`
   and `bulk_message_deletion`, and property `include_subscribers`,
   control various details that are expressed in the types in
   `eventTypes.js` and `modelTypes.js`.

So the values we pass here should be determined inside the API
bindings, not presented as part of the interface that a caller of
the bindings can manipulate at will, because in fact it can't
without breaking other aspects of the bindings.