android notif [nfc]: Update a TODO comment with issue link
I just filed this issue for this task today.

Also indent the continuation line of the TODO.  I find that a
helpful convention just for visually parsing the comment...
plus in Android Studio it turns out that the IDE recognizes
TODO comments and can give them a distinct color, and it expects
this convention too.