android notif: Explain URL we generate and ignore; fix rare bug
This URL generation dates to 2004dcf40 in 2018-12.  The intended
Android idiom is that the URL should on its own identify what the
thing is that the Intent expresses an intention to open or view.

We don't actually use it for that, though; we just read the "extra"
data.  So I just tried cutting it out in favor of `null`.  That works
fine for opening one notification -- `Intent` doesn't require the URL
be non-null, and in fact it has a constructor that takes just an
action string and leaves the URL null, and the getter `getData` for
the URL is marked as `@Nullable`.

But when you try to open a second notification, it takes you to the
same place as the first one did.

Same result if I make the URL a constant like `zulip://open-notif`.
Seems like something is taking just the action and the URL and so on,
without the extras, and expecting that to uniquely identify the
Intent; even perhaps caching what the extras should be given those.

But then that's not only an obstacle to simplifying this, which one
would simply want to turn into a comment.  It actually points to a bug
in the existing code, though one that requires quite a coincidence to
trigger: if you have two different notifications *with the same Zulip
message ID* -- so in particular on different servers, where different
messages that notified you just happened to collide on their message
IDs -- and you open one, then open the other, the app will behave as
if you opened the first one again.

So, fix that bug, and explain why this code isn't simpler.