android notif [nfc]: Comment on changing channel's settings
Each of these settings is here because it had been in our old code
for creating the notification itself.  When we originally started
creating a notification channel, in commit 2f7d1f657 in 2018-10,
we neglected to carry over these settings to the channel; which meant
that the settings had no effect on Android 8 Oreo and later.  That
means they've had no effect for an increasingly-overwhelming majority
of our Android users, up until we started setting them on the channel
in the last few months.

Setting them on the channel, in turn, has no effect for existing
installs.  I'd known that but it doesn't look like we discussed it
when making these changes recently, so I must have forgotten about it.
It's certainly easy to miss -- we're used to changing the code being
enough to change how it behaves -- so this function deserves a comment
explaining that situation.

On the other hand... all of these sure look like they're setting the
values that would be the defaults anyway.  That is, it seems like
they were probably redundant in the old `getNotificationBuilder` in
the first place, and remain redundant here.  In that case all is
well, but we should definitely have a comment to bear this in mind
for future changes we want to make to these notification settings.

Also add a TODO about perhaps devising a nice distinct vibration
pattern and then setting that.  This was in the old code, dating from
458ef8832 in 2019-01.  We never did act on it, but with our spiffy
new notifications and cleaner code here, perhaps now is the time.