ios notif: Validate type of user_id field, when present
As we're already doing on Android; see extractIdentity in
FcmMessage.kt.

Next, we'll expose this data from iOS and Android notification
payloads, when present, to inner app code, by adding
`user_id?: UserId` to our Notification type.

This brings the implementation in line with the comment at the top
of the `extract` module that says that servers have started
including a user_id.