android notif: Set a distinct vibration pattern
For detailed discussion of choosing this particular pattern, see:
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/notification.20vibration.20pattern/near/1284530

This also makes the first time we're discarding an old notification
channel / category and making a new one, in order to change the
settings that can't be changed on an existing channel.  So add a bit
of infrastructure to handle that: just look for channels that aren't
at the current channel ID, and delete them.

Before putting this in a prod release, I'd like to bunch it up with
setting a distinct notification sound too, to avoid repeatedly
clobbering users' settings.  But it's fine to do that for us few
alpha users, so that isn't an obstacle to merging this now.

Also delete the comments about settings being added recently and not
necessarily applying, because with the new channel ID they will now
always have been applied.

Fixes: #5136