persist [nfc]: Drop parameters to persist-complete callback
We already weren't using these parameters; we use the callback only
to mark timing.

There isn't an obvious use case for the `restoredState` parameter
that wouldn't be a bug.  By the time the callback is invoked with any
non-void value there, we've already dispatched the rehydrate action,
so the callback could simply use `store.getState()` if it wants to
inspect the new state.  And if it were to instead look at the partial
state read from storage... it's not clear what valid purpose that
would serve.

As for the `err` parameter, the most reasonable use for that would be
for logging.  But when there's an error, we already log from inside
`getStoredState`, and do so in more detail.