startEventPolling [nfc]: Be clearer that this is a per-account action
In a multi-account, post-#5005 world, we're likely to want a live
event queue for all the logged-in accounts. So we'll want this
function to start polling for "this account", not for the global
"active account". The action is already typed as a `ThunkAction`,
not a `GlobalThunkAction`; that's a good start.

The `auth` object we're working with after the code change is the
same one as before, we just get it more directly. (`tryGetAuth` uses
`getAccount`, which, pending #5005, assumes "the account" is the
active account; see the implementation comment in `getAccount`.)
And, with #5005, it will naturally change to be the Auth object for
"this account", the one the caller cares about.