msglist: Show new action sheet on long-pressing a PM recipient header
Instead of a toast with the recipients' names.

When we start targeting Android 12, which we hope to do soon, toasts
will be limited to two lines [1]. It was never a great design to use
a toast here in the first place, since in theory you can have a
group PM conversation with arbitrarily many members, and there's
only so big a toast can get before it starts to look weird or cut
things off in any implementation. With just two lines of space, the
toasts will be even less helpful.

Better to give an action sheet, like we do on long-pressing topic
headers. The action sheet gives a button to take you to the PM
recipients screen, where you can see all the details of the
recipients.

[1] https://developer.android.com/about/versions/12/behavior-changes-12#toast-redesign

Fixes-part-of: #5171