lightbox: Improve share-failed error messages, and comment on one
The "Can not download" message in particular is misleading -- the
user didn't press a button to download the image, but to share it.
So, change that message to be the same as the one above it (make
them both "Share failed"). Then cut out the inner try/catch because
the outer one causes the same behavior.

The comment is adapted from Greg's GitHub comment at
  https://github.com/zulip/zulip-mobile/pull/5175#discussion_r773452105