messageListElementHtml [nfc]: Only do one element; make callers map/join
Since the `messageListElementHtml` tests aim to cover HTML for
entire message lists, not their pieces individually -- they generate
snapshots using the map/join logic -- there's a more appropriate way
to think of them now: as tests of the `updateContent` function.
That's basically what they are, they're just missing coverage of
details like `scrollStrategy`.

It turns out that `scrollStrategy` has some tests already, and
there's one very simple test of `updateContent` with the following
name:

  'when the rendered messages differ (even deeply) a "content"
    message is returned'

Normally we'd go ahead and audit for full coverage of the
`updateContent` function. But that's about to be made obsolete by a
new "edit-sequence" event type. Those "edit-sequence" events will be
constructed with help from a new message-list diffing algorithm
(for #3851).

Since this "content" event type and the new edit-sequence event type
both control how message lists are rendered, we'll substantially
reuse the code in this file to test the latter.