msglist: Implement "message-list diffing" algorithm!
We don't use it yet because we're waiting for more complete test
coverage.

We start by verifying that the algorithm can give the same whole
message lists as the old "update content" algorithm we currently use
in production. To do that, we actually steal almost all of the test
code from there, and check that the snapshots still match.

That's not quite enough for full coverage, though. It only exercises
a special case of the new algorithm: an edit sequence that takes the
message list from empty to full with just a lot of inserts. The
algorithm has a lot of other interesting cases we should test, like
"delete" and "replace" instructions.

For more test coverage, we'll make a new "check" function to run on
various combinations of the same messages. It'll take an old and a
new $ReadOnlyArray<Message | Outbox>, ask our algorithm to produce
an edit sequence, and check that the edit sequence actually takes
you from the old message list to the new one [1].

[1] To construct the starting point (old message list) and the
    goalpost (new message list), we'll do the same thing we've done
    here: simulate changing a list from "empty" to "full", with an
    edit sequence that's inevitably just a lot of inserts. That
    might be pretty bad test design (what if there's a bug in how
    inserts work?), except that this commit exhaustively
    snapshot-tests exactly those kinds of constructions, with the
    same messages, so they should be pretty reliable.

Related: #3851