narrow: Store stream IDs, not only names!
In the migration tests, it is kind of messy that we have to update
a previous migration's test.  That happens because we're running
all the migrations to completion.

As the comment at the "whole sequence" test says about a related
point: this is probably not a great design for continuing to add
more migrations in this sequence.  But pretty soon we're going to
cap it off and write future migrations in a new framework, which
will come with its own way of writing tests.  So for the moment
just live with it.

Fixes-partly: #4333