getMessageListElements tests: Remove not-so-helpful `deepFreeze` calls
I'm not sure we've ever caught a bug using `deepFreeze` like this. I
think such a bug would come with a pretty strong code smell that
we'd catch in review.

We've also gotten better about marking our Flow types read-only, and
that makes the `deepFreeze`s even more likely to be redundant.