unread: Properly handle update_message when stream but not topic edited
From the documentation update in zulip/zulip#20693 we learn (among
other things) that in this situation `orig_subject` is indeed present,
but `subject` isn't.  We'd been implicitly assuming the reverse: that
`subject` would be present for any stream or topic move, even if not
`orig_subject`.  (And our types were saying it's *always* present,
even for pure content edits, even of PMs, which was clearly wrong.)

Update our types and our tests to reflect that, as well as our code
in the unreads reducer.  The new version of the "if stream updated"
test successfully fails on the old version of the reducer.