migration types: Move a FlowFixMe inside migrations list
Inside these migration functions, we pretend that the object they're
passed is really of type GlobalState, with all the properties and the
latest types for those properties.  But that's a lie; it wouldn't be
much of a migration system if it were true.

Make the type exported from this module slightly less of a lie, so
that some of the lying is done inside the module.

This has the effect of decoupling the lie told inside the module from
the one told on the outside.  That will let us go on to adjust exactly
what lie is told inside this module, without having the new type
propagate out to where the migrations are invoked.