narrow: Stop storing stream names!
This is nearly NFC, because we've already eliminated all the places
that we actually consumed these stream names... other than in
`keyFromNarrow`.  Those keys are used not only to reconstitute Narrow
objects (in `parseNarrow`) but also to uniquely identify a narrow in
our data structures.  So there is a subtle bug this fixes: if a stream
gets renamed, then before this change we could get duplicate data for
it, and conversely could fail to find an old draft.

We leave a few things to clean up later in larger, boring, NFC
commits, in order to keep this interesting change tightly focused:

 * The constructors streamNarrow and topicNarrow still accept a
   "stream name" argument, but they ignore it and allow it to be void.

 * The streamNameOfNarrow function still exists, but returns void.
   All its call sites have already been eliminated, other than for
   passing the result directly back to a Narrow constructor.

 * The `caseNarrow*` functions still pass a "stream name" argument
   to the `stream` and `topic` callbacks, but it's always void.
   All callers have already switched to ignoring that argument in
   favor of the stream ID, other than the few we switch over in
   this commit.

Fixes: #4333