android notif: Skip creating channel if already exists
If there's already a notification channel with the ID we'd use (i.e.,
if this isn't the first launch after install, or after upgrade from a
version that used a previous channel ID), then trying to create the
channel would do nothing anyway.  So skip trying.

This saves (a negligible amount of) work where we were setting up the
NotificationChannel object describing the channel we'd like to create.
More usefully, when we start supplying notification sounds, we'll
start doing a bit more work at channel creation time, to copy our
sound files into shared storage.  Even looking to see if the files are
already there can take on the order of 30ms, so it's good to not have
to do that on every launch.

This function still takes about 3ms to run on my Pixel 4 (with
Android 11) when there's nothing to do.  Not awesome, but on the
scale of our performance concerns that's quite tolerable.  And if
that's just how long it takes to get ahold of the NotificationManager
service and find out if we already have our notification channel,
then it may be basically up to Android and hard for us to do better.