emoji: Add two Wordle emoji names to unicodeCodeByName
This is a tactical fix, corresponding to zulip/zulip#21023
and zulip/zulip#21031. Now that the server is recognizing these
emoji, and turning them from Unicode into Zulip emoji objects,
they come out broken unless we have them in our client-side
database of Zulip emoji.

We should really start asking the server to provide this data
instead of hard-coding it ourselves. That's #2956.

But, as Greg says at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Emoji.20sync/near/1322629 :

> A tactical fix for Wordle is probably about two lines, so I think
> that's worth doing. This is the one time in several years that
> there's been such a sudden surge in the popularity of a handful of
> specific emoji.

And to do #2956 most correctly, it'll be best to have the server
document how it provides the relevant data. I've just filed
zulip/zulip#21037 for that.

For a screenshot of how Wordle was broken before this tactical fix,
see
  https://chat.zulip.org/#narrow/stream/97-off-topic/topic/Wordle/near/1322607 .