user status model: Use new selector everywhere outside model files
To get these accesses encapsulated inside the code responsible for
this part of the model.

I enlisted Flow to help me find all the places where we weren't
using the getter: I stuck a `| void` at the end of the
`UserStatusState` definition, and went through the complaints of
accessing a property on possibly-undefined.