msglist tests: Add HTML snapshot tests for interesting single messages
As suggested by Greg at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/msglist-diffing.20tests/near/1320910

In particular:

- reactions
- polls
- starred messages (and all other flags in FlagsState)
- muted senders

I've gone ahead and removed the

  // TODO: Test with a variety of different things in
  // `backgroundData`.

-- not because I've thought hard enough to say we've exhausted every
possible thing we'd eventually want to test, but because this seems
like decent progress, and definitely a "variety" of things. :-) As
we do with all our tests, we'll naturally add more cases as we think
of them.