flow: Apply our Object.freeze workaround in savePushToken
Flow 0.149 would still treat {} as inexact (because "unsealed").

From there, it would rightly complain that spreading an inexact
`extraParams` could end up clobbering the `token` property with some
non-string value. So that's an improvement in soundness -- it's just
that in this case, as in the others where we've applied this
workaround, we don't want the {} to be treated as inexact.