reactions [nfc]: Cut a long-disabled style tweak, on borderWidth
We've had this `borderWidth: 0.15` line here since this screen was
introduced, in a49b1f1f8 in 2019-11.

But then since 427a40eb1 in 2020-10, it's been overridden by the
`borderWidth: 0` in `materialTopTabNavigatorConfig`, so it's had
no effect.  And checking now, this screen looks fine.

And it's not clear why we would want a border to be different in the
tabs on this screen vs., say, the tabs on the main "streams" screen.
Probably we don't.

So cut that out, and simplify away that customization hook.