mute [nfc]: Add some TODO comments for muting users
In the previous commit, we learned we weren't updating an existing
message list on an event saying that a user was muted or unmuted.

This had gone unnoticed in part because we don't have good
systematic tests for this logic in the existing code; we
discovered it on writing tests for the new message-list-diffing
logic.  (Thanks, tests!)

The other reason this had gone unnoticed is that it actually takes
a bit of a coincidence for you to have a message list open while
muting or unmuting someone -- because we don't actually have any
UI for doing so in the mobile app!  There's one existing TODO
comment related to that; give it a reference to the issue for this
feature, #4655.