ChatScreen: Schedule fetch if list suddenly empty
There's a bug that's currently somewhat hard to trigger: if you're
looking at some narrow, and then all the messages that we have in that
narrow disappear, we'll switch to saying "No messages" -- and won't
attempt any fetching of more messages -- even when the truth is that
we just don't know if there are any messages.  The user ends up having
to navigate out and come back, and only then do we go fetch messages.

This occurs even though our `state.caughtUp` will correctly say that
we don't know if there are any messages.

Currently this can happen if all the messages are deleted (perhaps
you're looking at a short conversation), or if you're looking at the
starred-messages or @-mentions narrow and the messages all lose the
respective flag.  As we start supporting moving a message / editing
its topic and stream, though, this will become more common.

In this fix, we make sure that at least we start fetching messages
when this happens.  The fetching will also replace the "No messages"
with placeholders showing we're fetching.

There may still be a flash of the untrue "No messages" message, but
only for a frame or two while the component rerenders -- it doesn't
wait for the fetch request, let alone stay stuck there indefinitely.