api: Update Message type on topic_links; stop maintaining subject_links
The two migrations here (of the property's name, and its type when
present) are parallel to migrations in the `update_message` event.

Ever since the rename (back in Zulip Server 3.0), we've been getting
message objects with a `topic_links` and no `subject_links`; and then
if the message is ever updated we add `subject_links: undefined` to
it.  The reason this API change hasn't surfaced as a bug is that we
never actually consume this property (under either name).

For the moment, just cut out the broken code rather than update it.
Properly handling the new situation will be more than a trivial
change, because we'll want to canonicalize how we store this property
internally.