migration: Preserve original list of keys for migration 24
This needs to be defined separately from the current list for the
same reasons as historicalStoreKeys.

Unlike the latter, we don't expect to keep making more migrations
that do the same thing (not before switching to sound storage and
the new migration framework), so there's no need to give this a
generic name -- we can just make it specifically about that one past
migration.  Then that also means there's no need for a test to check
that the current value still agrees.

This change is not NFC because... we actually had changed this list
since creating this migration!  See 1b2567cfe:src/boot/store.js for
the original list, and 14ee13e1a for the change.  We didn't notice
that that change had the effect of altering this past migration's
behavior, oops.

Fortunately the change was harmless -- it just caused us to purge
an extra key, one that wouldn't be present anyway.  But it nicely
illustrates why this needs a separate definition.