UserStatusScreen [nfc]: Make statusText state a string
UserStatusScreen has undertaken to use Input as a controlled input,
where at all times it's telling the Input what `value` it should
have. https://reactjs.org/docs/forms.html#controlled-components

It seems like the responsible thing for UserStatusScreen to do, for
clarity, is to at least store the value meant for `value` in a
proper state variable. We'll solidify that role with a rename,
coming next.