AsyncStorage [nfc]: Write some jsdoc
Hmm, somehow I'd thought that the upstream version was unsound
even on Android when it comes to multiSet!  But, looking at the
implementation now, it sure seems to use a single transaction; and
that's true already in the initial release from 2015.

Anyway, this gives us soundness on both platforms, which is a big
relief and opens up opportunities for more fearless improvements to
our schema.  It also gives us direct control over the database, which
we'll use for migrations and to get more flexibility in the schema.