api [nfc]: Remove some param-defaulting within our api.createStream binding
I think the only good reason to have such fallback code at this
layer would be that we don't trust the server to use reasonable
fallbacks itself.

There are certainly awkward things about this endpoint -- it's doing
double duty as "subscribe-to-stream" and "create-a-stream" -- but I
see no reason to distrust the server's defaults for these params,
which set attributes on a created stream. Its choices currently
match ours, but they could reasonably change, and we'd be fine
deferring to the server if they did.

NFC because, currently, this binding's only caller also applies the
same fallbacks. That might be a bad choice, but at least now that
choice isn't forced by the code that's designed to interface
directly with the server.