ComposeBox [nfc]: Shrink a useCallback dependency list
This doesn't actually use the value of `state`, so there's no need
for it to depend on it.

If we simply delete `state` from the list, then the ESLint rule
react-hooks/exhaustive-deps complains.  The reason seems to be the
`typeof state` in the type annotation -- it seems the rule has a bug
where it looks inside type expressions, not only value expressions,
when looking for references to possible dependencies.  Just delete
the annotation; Flow infers the type just fine.

(If we did need the annotation, e.g. if we were exporting
`useUncontrolledInput`, I expect we could solve the problem by
naming that type with a type alias.)