deps: Upgrade to Node v16, in setup docs and in CI
I've been using this for local development for a few weeks now,
and all seems well.  (That's longer than it needs to be; I just
hadn't come back to this to follow up.)

Node v12 is going EOL in just over a week, 2022-04-30, so it's
past time to make this upgrade official.

Also delete an old troubleshooting entry for a problem with Node v11
that was avoided by sticking to v10; and fix two other references to
Node v10 (where the problem was using an older version like v8) so
that they remain accurate now and for further upgrades in the future.

In particular, where we had been checking in our tests for Node v10+
(but in fact for >=10 and <20), just replace it with handy metadata
in our package.json to make Yarn (or `npm`) complain at install time.
Thanks to Anders Kaseorg for pointing out that package.json feature.

Fixes: #4263