initial data types [nfc]: Put enter_sends under update_display_settings
(And remove the TODO comment, because there's already one on the
whole of InitialDataUpdateDisplaySettings.)

This is where it belongs. Oddly, the doc lists this property among
many that come with realm_user, which I guess is why it ended up in
InitialDataRealmUser.. But the doc still says "Present if
`update_display_settings` is present in `fetch_event_types` […]" for
this one:
  https://zulip.com/api/register-queue
, which is what we'd expect, given that this is one of the values
that's migrated to the new `user_settings` object; see #4933.