api types: Add UserTopic type, preparing for user_topic API
This just adds type definitions, without changing the types we
expect from the server.

After this, as the next steps:

 * we'll migrate the data structure we maintain in our own store to
   one that can smoothly handle the new API format;

 * then as a separate step after that, we'll switch to actually
   consuming the new API format.

It's convenient to use this UserTopicVisibilityPolicy type in
describing that new data structure's type.  Hence we add it early,
in this commit, before we actually start requesting the new format
(or update our event and initial-data types to reflect doing so).