types: Switch to TsFlower for react-navigation types
This replaces the type definitions we've had for react-navigation,
which derive from the flow-typed project with some changes and fixes
of our own, with new definitions based more directly on the
TypeScript type definitions found in the library itself upstream.

---

The Flow type definitions are first generated from the upstream TS
type definitions using TsFlower, a tool I recently wrote for this
purpose:
  https://www.npmjs.com/package/tsflower

The output is auto-formatted, and then a stack of local patches
is applied.  There are 24 of these, including one added for
react-native-safe-area-context in the previous PR.  They comprise:

 * 9 patches for things that could be fixed upstream, e.g. by adding
   `readonly`.  These are places where the TS definitions are less
   accurate than they could be, but TS apparently takes the
   distinctions less seriously than Flow does.

 * 8 patches for things that could be fixed in TsFlower.

 * 5 "irreducible" patches, where we fundamentally need a local
   patch like this.

   * 3 patches are because TS doesn't have a way to express a
     distinction that becomes important in Flow: exact object types,
     or covariant type aliases.

   * 2 patches are because the types use a TS feature that there's
     no way to translate into Flow: conditional types.  As a result,
     our patched-in versions are imperfect approximations, but I
     think good enough.

 * 1 patch that just reformats a few definitions for readability.
   This was useful in developing some of these changes.

 * 1 patch (the one at the end) that fixes an issue I haven't fully
   debugged.

The best way to read these patches is probably to use the command:

    $ tools/tsflower unpack -s

in order to unpack them into a Git branch, and then read that branch
in the usual way.  (The `-s` just makes the command faster, by not
rerunning TsFlower.)

---

Meanwhile, because these upstream-based type definitions are quite
different from the ones we had based on flow-typed, our own code
that interacts with react-navigation needs a number of small changes.

All of these changes are purely in the types, with no runtime
changes -- after all, we're using the same underlying library JS
code at runtime, just switching from one type-level description of
its API to a different one.

I pulled out a number of such changes as prep commits in the branch
leading up to this one, for changes that worked already with the old
definitions.  The changes in this commit, happening simultaneously
with the change of type definitions, include:

 * Some fixmes and expected errors get a new error code.

 * Two fixmes in EditStreamCard go away.

 * New, quite appropriate, fixmes in AppNavigator.

 * Regression in a type-test: we no longer get an error in
   `<Foo.Screen …>` if the `name` prop doesn't agree with the name
   in the component's nav props.  Though perhaps not that big a
   regression in practice, because the error you'd get if you made
   this mistake in the app code was a baffling one:
     https://github.com/zulip/zulip-mobile/pull/5397#discussion_r887353922

 * The new type of `useNavigation` is pretty useless.  But we already
   have a type-wrapper anyway; just update it.

 * The type StackAction is now called StackActionType.  The old name
   seems better (values of that type are stack actions, not types of
   stack actions); but so it goes.

 * NavigationState now takes type arguments; the defaults are fine,
   but we need an explicit `<>`.

 * The createFooNavigator functions now take just one type argument,
   not three; it corresponds to what was the first one.