user: When viewing own profile, show the same info others see
There are two different screens where we show a variety of profile
data:

 * When you navigate to a user's profile, we show AccountDetailsScreen.

 * When you go to the rightmost tab in the app's main screen, the one
   with your avatar as the icon, we show ProfileScreen.

The names of these components aren't real helpful.  But we naturally
want different combinations of information on them -- in particular
ProfileScreen is where one finds several pieces of UI that some users
will want to use regularly, so it's important that it not get too
cluttered.

The two screens use a shared component AccountDetails for some of the
information they both want to display.  Then for some pieces of
information that we want to show on AccountDetailsScreen but not on
ProfileScreen, the way we'd implemented that is by having
AccountDetails go ahead and show them -- but only if the user we're
looking at is not the self-user.

That has the side effect that if you take the trouble to go navigate
to your own profile, say by visiting the self-1:1 thread and hitting
the "info" icon, we still won't show you your full profile in the way
it shows up for other people.  That's unhelpful.

Instead, have AccountDetails show the same information for any user.
For things we want to leave out of ProfileScreen, leave them out of
AccountDetails, and have AccountDetailsScreen handle them.

This commit should make no change on ProfileScreen, and no change on
AccountDetailsScreen when viewing any profile but your own.  The
styles added to the wrapper views mimic what the ComponentList helper
(used by AccountDetails) was doing with them.