nav [nfc]: Use navigation.push instead of navigation.dispatch
This replaces all the call sites of the form
`navigation.dispatch(navigateToFoo(…))`.  We effectively inline the
definitions of the respective `navigateToFoo` nav-action creators,
to directly invoke `push` on the navigation prop.

A nice feature of this change is that while the definitions of these
nav-action creators in `navActions.js` have no effective type-checking
-- in particular, there's no check that the given route params
correspond to the types the given route expects -- these `push` calls
do get such type-checking, because these `navigation` objects have
appropriate types.

This commit doesn't cover all the call sites of these `navigateToFoo`.
Where we're dispatching the result on `NavigationService` instead of
on a component's `navigation` prop (whether gotten directly, passed
down from a parent component, or gotten equivalently through
`useNavigation`), there is no `push` method, so in those places we
stick for now with `dispatch`.