nav [nfc]: Cut now-disused navActions action creators
This covers most of the `navigateToFoo` nav-action creators, but
sadly not all of them: as discussed in the previous commit where we
simplified away most of the call sites, there are still some that
remain in places where we still use NavigationService.  So for now
we keep the nav-action creators that are used there.

Fixes: #4417